feat: add eino assistant as example, update files
