--- quickstart/eino_assistant/agent/agent_test.go
@@ -1,5 +1,0 @@
-package agent
-
-// func TestRun(t *testing.T) {
-// 	RunTestRAG()
-// }
