+++ quickstart/eino_assistant/cmd/server/agent/web/style.css
@@ -1,6 +1,5 @@
 body {
     margin: 0;
-    padding: 20px;
     font-family: Arial, sans-serif;
     background-color: #f5f5f5;
 }
