+++ quickstart/eino_assistant/data/memory/ce10d8a8-a866-474f-919a-c1ef094c8236.jsonl
@@ -1,5 +1,8 @@
 {"role":"user","content":"你好，可以帮我介绍一下 Eino 的优势吗？ Thanks♪(･ω･)ﾉ"}
-{"role":"assistant","content":"Eino具有以下优势：\n- **借鉴优秀框架且符合Go语言习惯**：Eino参考了开源社区中诸多优秀的AI应用开发框架，例如LangChain、LangGraph、LlamaIndex 等，提供了更符合Golang编程习惯的AI应用开发框架。 \n- **编排优势明显**：以组件为第一编排对象，同时提供具有极强扩展能力的Lambda节点作为编排对象，能够实现快速上手和定制扩展的双优势。并且编排过程中强化了 “数据流” ，callbacks 提供了观测和调试的基础能力。 \n- **助力项目精准定义与成本控制**：其AI处理功能可帮助用户准确地定义项目，进而降低部署成本。\n- **响应灵活**：能快速响应不断变化的客户或项目需求，让用户对所做更改充满信心。 \n- **支持团队协作**：支持团队成员在单个项目上进行协作，并可通过编辑历史和项目状态进行管理。 \n\n你可以点击[此处](https://www.cloudwego.io/zh/docs/eino/overview/)查看更多相关信息。  ","response_meta":{"finish_reason":"stop","usage":{"prompt_tokens":1973,"completion_tokens":234,"total_tokens":2207}}}{"role":"user","content":"Eino 的项目地址是什么？官方文档在哪里？可以帮我打开下吗"}
+{"role":"assistant","content":"Eino具有以下优势：\n- **借鉴优秀框架且符合Go语言习惯**：Eino参考了开源社区中诸多优秀的AI应用开发框架，例如LangChain、LangGraph、LlamaIndex 等，提供了更符合Golang编程习惯的AI应用开发框架。 \n- **编排优势明显**：以组件为第一编排对象，同时提供具有极强扩展能力的Lambda节点作为编排对象，能够实现快速上手和定制扩展的双优势。并且编排过程中强化了 “数据流” ，callbacks 提供了观测和调试的基础能力。 \n- **助力项目精准定义与成本控制**：其AI处理功能可帮助用户准确地定义项目，进而降低部署成本。\n- **响应灵活**：能快速响应不断变化的客户或项目需求，让用户对所做更改充满信心。 \n- **支持团队协作**：支持团队成员在单个项目上进行协作，并可通过编辑历史和项目状态进行管理。 \n\n你可以点击[此处](https://www.cloudwego.io/zh/docs/eino/overview/)查看更多相关信息。  ","response_meta":{"finish_reason":"stop","usage":{"prompt_tokens":1973,"completion_tokens":234,"total_tokens":2207}}}
+{"role":"user","content":"Eino 的项目地址是什么？官方文档在哪里？可以帮我打开下吗"}
 {"role":"assistant","content":"Eino 的官方文档地址为：https://www.cloudwego.io/zh/docs/eino/ ，我已经帮你打开了该文档链接。如果你需要了解 Eino 项目地址等更多信息，可以在文档中进一步查找。如果你还有其他需求，请随时告诉我。  ","response_meta":{"finish_reason":"stop","usage":{"prompt_tokens":1102,"completion_tokens":59,"total_tokens":1161}}}
-{"role":"user","content":"帮我定一个 2025-1-17 19:00 的 todo，要把 Eino 的 example 通读一遍"}
+{"role":"user","content":"在网上查一下 cloudwego 有什么新的项目发布吗？"}
+{"role":"assistant","content":"为你查询到CloudWeGo的相关信息如下：\n- CloudWeGo是一套由字节跳动开源的、可快速构建企业级云原生微服务架构的中间件集合。该项目共同的特点是高性能、高扩展性、高可靠，专注于微服务通信与治理。\n- 第一批开源了四个项目：Kitex、Netpoll、Thriftgo和netpoll-http2， 以RPC框架Kitex和网络库Netpoll为主。最近又发布了面向 AI 应用场景的开发框架 Eino，致力于让开发者轻松构建高质量的 AI 应用。\n\n更多信息可查看[CloudWeGo官网](https://www.cloudwego.io/)。","response_meta":{"finish_reason":"stop","usage":{"prompt_tokens":2377,"completion_tokens":186,"total_tokens":2563}}}
+{"role":"user","content":"帮我定一个大后天晚上 19:00 的 todo，要把 Eino 的 example 通读一遍"}
 {"role":"assistant","content":"已成功为你添加一个待办事项：在2025-1-17 19:00，把 Eino 的 example 通读一遍。","response_meta":{"finish_reason":"stop","usage":{"prompt_tokens":1083,"completion_tokens":36,"total_tokens":1119}}}
