+++ quickstart/eino_assistant/data/todo/todos.jsonl
@@ -1,1 +1,2 @@
 {"id":"a13f1db5-0a00-4a9f-8151-52fdc097b11a","title":"阅读 Eino example","content":"把 Eino 的 example 通读一遍","completed":false,"deadline":"2025-1-17 19:00:00","is_deleted":false,"created_at":"2025-01-13T22:45:13+08:00"}
+{"id":"cc41377b-c6b0-4521-a81b-f0eca0be5459","title":"读书计划","content":"读书","completed":false,"deadline":"2025-01-14 22:00:00","is_deleted":true,"created_at":"2025-01-13T23:28:56+08:00"}
